Added Sqlite Commands to the CommandRunner to get a list of quote lookup requests, and add a quote lookup response to the table. Also made it so when a company is added to the lookup table, a quote history table is created for that company.

+++ NeuralStocks/ApiCommunication/QuoteLookupResponse.cs
@@ -5,34 +5,34 @@
         public string Status { get; set; }
         public string Name { get; set; }
         public string Symbol { get; set; }
-        public float LastPrice { get; set; }
-        public float Change { get; set; }
-        public float ChangePercent { get; set; }
+        public double LastPrice { get; set; }
+        public double Change { get; set; }
+        public double ChangePercent { get; set; }
         public string Timestamp { get; set; }
-        public float MarketCap { get; set; }
-        public float Volume { get; set; }
-        public float ChangeYtd { get; set; }
-        public float ChangePercentYtd { get; set; }
-        public float High { get; set; }
-        public float Low { get; set; }
-        public float Open { get; set; }
+        public double MarketCap { get; set; }
+        public double Volume { get; set; }
+        public double ChangeYtd { get; set; }
+        public double ChangePercentYtd { get; set; }
+        public double High { get; set; }
+        public double Low { get; set; }
+        public double Open { get; set; }
 
         public QuoteLookupResponse()
         {
             Status = "";
             Name = "";
             Symbol = "";
-            LastPrice = 0f;
-            Change = 0f;
-            ChangePercent = 0f;
+            LastPrice = 0;
+            Change = 0;
+            ChangePercent = 0;
             Timestamp = "";
-            MarketCap = 0f;
-            Volume = 0f;
-            ChangeYtd = 0f;
-            ChangePercentYtd = 0f;
-            High = 0f;
-            Low = 0f;
-            Open = 0f;
+            MarketCap = 0;
+            Volume = 0;
+            ChangeYtd = 0;
+            ChangePercentYtd = 0;
+            High = 0;
+            Low = 0;
+            Open = 0;
         }
     }
 }
